Move connection status to connect-status-hooks.tsx and document each status and what they mean.
